@@ -6,16 +6,19 @@
 	"files": [
 		"build",
 		"CHANGELOG.md",
+		"LICENSE",
 		"README.md"
 	],
 	"types": "build/index.d.ts",
 	"description": "Reusable ReactJS Component Library",
 	"scripts": {
 		"build": "rollup -c",
+		"start": "rollup -c -w",
 		"test": "jest",
 		"test:watch": "jest --watch",
-		"generate": "node ./commands/create-component",
-		"generate:forwardRef": "node ./commands/create-forwardRef-component",
+		"mkc": "node ./commands/create-component",
+		"mkfc": "node ./commands/create-forwardRef-component",
+		"mkcon": "node ./commands/create-context",
 		"prepublishOnly": "npm run build"
 	},
 	"repository": {
@@ -36,31 +39,30 @@
 	"homepage": "https://github.com/Dodobrat/react-ui-kit#readme",
 	"deprecated": false,
 	"peerDependencies": {
-		"react": ">=16.8.0",
-		"react-dom": ">=16.8.0"
+		"react": "^16.8.0",
+		"react-dom": "^16.8.0"
 	},
 	"devDependencies": {
-		"@rollup/plugin-commonjs": "^11.1.0",
-		"@rollup/plugin-node-resolve": "^7.1.3",
-		"@testing-library/jest-dom": "^5.5.0",
-		"@testing-library/react": "^10.0.2",
+		"@rollup/plugin-commonjs": "^17.1.0",
+		"@rollup/plugin-node-resolve": "^11.2.0",
+		"@testing-library/jest-dom": "^5.11.9",
+		"@testing-library/react": "^11.2.5",
 		"@types/classnames": "^2.2.11",
-		"@types/jest": "^24.0.24",
-		"@types/react": "^16.9.12",
-		"@types/react-dom": "^16.9.8",
+		"@types/jest": "^26.0.20",
+		"@types/react": "^17.0.2",
+		"@types/react-dom": "^17.0.1",
+		"babel-core": "^6.26.3",
+		"babel-runtime": "^6.26.0",
 		"classnames": "^2.2.6",
-		"focus-trap-react": "^8.3.0",
-		"identity-obj-proxy": "^3.0.0",
-		"jest": "^24.9.0",
-		"node-sass": "^5.0.0",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
-		"rollup": "^1.27.4",
-		"rollup-plugin-copy": "^3.3.0",
-		"rollup-plugin-peer-deps-external": "^2.2.0",
-		"rollup-plugin-postcss": "^3.1.2",
-		"rollup-plugin-typescript2": "^0.27.0",
-		"ts-jest": "^24.2.0",
-		"typescript": "^3.7.2"
+		"focus-trap-react": "^8.4.2",
+		"jest": "^26.6.3",
+		"react": "^17.0.1",
+		"react-dom": "^17.0.1",
+		"rollup": "^2.39.0",
+		"rollup-plugin-peer-deps-external": "^2.2.4",
+		"rollup-plugin-sass": "^1.2.2",
+		"rollup-plugin-typescript2": "^0.29.0",
+		"ts-jest": "^26.5.1",
+		"typescript": "^4.1.5"
 	}
 }
