@@ -1,39 +1,27 @@
 import peerDepsExternal from "rollup-plugin-peer-deps-external";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
+import sass from "rollup-plugin-sass";
 import typescript from "rollup-plugin-typescript2";
-import postcss from "rollup-plugin-postcss";
-import copy from "rollup-plugin-copy";
 
-const packageJson = require("./package.json");
+import pkg from "./package.json";
 
 export default {
 	input: "src/index.ts",
 	output: [
 		{
-			file: packageJson.main,
+			file: pkg.main,
 			format: "cjs",
+			exports: "named",
 			sourcemap: true,
+			strict: false,
 		},
 		{
-			file: packageJson.module,
+			file: pkg.module,
 			format: "esm",
 			sourcemap: true,
 		},
 	],
-	plugins: [
-		peerDepsExternal(),
-		resolve(),
-		commonjs(),
-		typescript({ useTsconfigDeclarationDir: true }),
-		postcss(),
-		copy({
-			targets: [
-				{
-					src: "src/assets",
-					dest: "build",
-				},
-			],
-		}),
-	],
+	plugins: [peerDepsExternal(), resolve(), commonjs(), sass({ insert: true }), typescript({ useTsconfigDeclarationDir: true })],
+	external: ["react", "react-dom"],
 };
